requirements.txt
fastapi
uvicorn[standard]
python-docx
docx2pdf
mammoth
pillow
pillow-heif
